+++ app/kumactl/cmd/apply/apply_test.go
@@ -38,7 +38,6 @@ var _ = Describe("kumactl apply", func() {
 	var rootCtx *kumactl_cmd.RootContext
 	var rootCmd *cobra.Command
 	var store core_store.ResourceStore
-
 	BeforeEach(func() {
 		rootCtx = &kumactl_cmd.RootContext{
 			Runtime: kumactl_cmd.RootRuntime{
@@ -374,7 +373,9 @@ var _ = Describe("kumactl apply", func() {
 
 		// then
 		Expect(buf.String()).To(Equal(
-			`mesh: default
+			`creationTime: "1970-01-01T05:30:00+05:30"
+mesh: default
+modificationTime: "1970-01-01T05:30:00+05:30"
 name: sample
 networking:
   address: 2.2.2.2