+++ app/kumactl/cmd/get/get_meshes_test.go
@@ -158,7 +158,6 @@ var _ = Describe("kumactl get meshes", func() {
 		var rootCmd *cobra.Command
 		var buf *bytes.Buffer
 		var store core_store.ResourceStore
-
 		BeforeEach(func() {
 			// setup
 			rootCtx = &kumactl_cmd.RootContext{