+++ app/kumactl/cmd/get/get_traffic_traces_test.go
@@ -74,7 +74,6 @@ var _ = Describe("kumactl get traffic-traces", func() {
 		var rootCmd *cobra.Command
 		var buf *bytes.Buffer
 		var store core_store.ResourceStore
-
 		BeforeEach(func() {
 			// setup
 			rootCtx = &kumactl_cmd.RootContext{