+++ app/kumactl/cmd/get/get_trafficpermissions_test.go
@@ -84,7 +84,6 @@ var _ = Describe("kumactl get traffic-permissions", func() {
 		var rootCmd *cobra.Command
 		var buf *bytes.Buffer
 		var store core_store.ResourceStore
-
 		BeforeEach(func() {
 			// setup
 			rootCtx = &kumactl_cmd.RootContext{