+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # CHANGELOG
 
 ## master
+* feat: expose CreationTime and modificationTime
+  [#677](https://github.com/Kong/kuma/pull/677)
 * feat: pagination in kumactl
   [#690](https://github.com/Kong/kuma/pull/690)
 * feat: secret management API